Cursor-Projects/crypto-market-watcher/requirements.txt
requests
pandas
python-telegram-bot
schedule 